+++ oneflow/core/ep/include/primitive/gather.h
@@ -29,9 +29,9 @@ class Gather : public Primitive {
   Gather() = default;
   ~Gather() = default;
   virtual void Launch(Stream* stream, const void* src, void* dst, const void* indice,
-                      const size_t num_indices, const size_t batch_dim_size,
-                      const size_t outer_dim_size, const size_t gather_dim_size,
-                      const size_t inner_dim_size) = 0;
+                      const int64_t num_indices, const int64_t batch_dim_size,
+                      const int64_t outer_dim_size, const int64_t gather_dim_size,
+                      const int64_t inner_dim_size) = 0;
 };
 class GatherFactory : public Factory<Gather> {
  public:
